fix: Restricted Country page loads just its modal and simbolic background
